Long time no see

+++ graphql/type-defs.ts
@@ -81,7 +81,7 @@ export default gql`
 
   type Episode {
     id: Int!
-    n: Int!
+    n: Float!
     thumbnail: String!
   }
